deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
pillow
numpy
scikit-learn==1.5.1